# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckInboxResponse": {
        "properties": {
          "inbox_message_count_estimate": {
            "title": "Inbox Message Count Estimate",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/InboxCheckStatus"
          }
        },
        "required": [
          "status",
          "inbox_message_count_estimate"
        ],
        "title": "CheckInboxResponse",
        "type": "object"
      },
...
      "EmailDraft": {
        "properties": {
          "body": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "draft_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Id"
          },
          "subject": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Subject",
            "type": "string"
          },
          "to": {
            "items": {
              "format": "email",
              "type": "string"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "To",
            "type": "array"
          }
        },
        "required": [
          "to",
          "subject",
          "body"
        ],
        "title": "EmailDraft",
        "type": "object"
      },
...
      "InboxCheckStatus": {
        "enum": [
          "checked"
        ],
        "title": "InboxCheckStatus",
        "type": "string"
      },
...
      "PublicErrorResponse": {
        "description": "Stable safe HTTP error envelope shared by all routes.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "detail"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
      "SaveDraftResponse": {
        "properties": {
          "draft_id": {
            "title": "Draft Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SaveDraftStatus"
          }
        },
        "required": [
          "id",
          "status",
          "draft_id"
        ],
        "title": "SaveDraftResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hexel Onebox Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/mail/check-inbox": {
      "post": {
        "operationId": "check_inbox_api_mail_check_inbox_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInboxResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Check Inbox Api",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/drafts": {
      "post": {
        "operationId": "save_draft_api_mail_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Draft Api",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails": {
      "get": {
        "operationId": "fetch_emails_mail_emails_get",
        "parameters": [
          {
            "description": "Folder: inbox, sent, spam, trash, starred, all",
            "in": "query",
            "name": "folder",
            "required": false,
            "schema": {
              "default": "inbox",
              "description": "Folder: inbox, sent, spam, trash, starred, all",
              "maxLength": 32,
              "minLength": 1,
              "title": "Folder",
              "type": "string"
            }
          },
          {
            "description": "Max number of emails per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Max number of emails per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Gmail pagination token",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Gmail pagination token",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch Emails",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}": {
      "delete": {
        "operationId": "delete_email_mail_emails__email_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Email",
        "tags": [
          "Email-Operations"
        ]
      },
      "get": {
        "operationId": "fetch_email_by_id_mail_emails__email_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Fetch Email By Id",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}/read": {
      "post": {
        "operationId": "mark_as_read_mail_emails__email_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark As Read",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}/restore": {
      "post": {
        "operationId": "restore_from_trash_mail_emails__email_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore From Trash",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}/star": {
      "post": {
        "description": "Set the desired star label exactly; no read-then-toggle race exists.",
        "operationId": "set_star_state_mail_emails__email_id__star_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StarStateUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Star State",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}/trash": {
      "post": {
        "operationId": "move_to_trash_mail_emails__email_id__trash_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Move To Trash",
        "tags": [
          "Email-Operations"
        ]
      }
    },
    "/mail/emails/{email_id}/unread": {
      "post": {
        "operationId": "mark_as_unread_mail_emails__email_id__unread_post",
        "parameters": [
          {
            "in": "path",
            "name": "email_id",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "Email Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request validation failed; the error and detail fields are safe for clients to display."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mark As Unread",
        "tags": [
          "Email-Operations"
        ]
      }
    },
...
  }
}